+++ game.py
@@ -1,0 +1,3 @@
+import xos
+
+xos.run_game("ball", web=False, react_native=False)
