+++ src/lib.rs
@@ -1,3 +1,14 @@
+use pyo3::prelude::*;
+use pyo3::pyfunction;
+use pyo3::pymodule;
+use pyo3::wrap_pyfunction;
+
+use std::process::Command;
+use std::{fs, thread};
+use std::time::Duration;
+use tiny_http::{Server, Response};
+use webbrowser;
+
 pub mod apps;
 pub mod engine;
 pub mod video;
@@ -22,4 +33,103 @@ pub fn start() -> Result<(), JsValue> {
     let game = option_env!("GAME_SELECTION").unwrap_or("ball");
     let app = apps::get_app(game).ok_or(JsValue::from_str("App not found"))?;
     engine::run_web(app)
+}
+
+fn build_wasm(game: &str) {
+    let mut command = Command::new("wasm-pack");
+    command
+        .env("GAME_SELECTION", game)
+        .args(["build", "--target", "web", "--out-dir", "static/pkg"]);
+
+    let status = command.status().expect("Failed to run wasm-pack");
+    if !status.success() {
+        panic!("WASM build failed");
+    }
+
+    println!("✅ WASM built to static/pkg/ with game: {}", game);
+}
+
+fn launch_browser() {
+    thread::spawn(|| {
+        thread::sleep(Duration::from_millis(500));
+        let _ = webbrowser::open("http://localhost:8080");
+    });
+}
+
+fn mime_type(path: &str) -> &'static str {
+    if path.ends_with(".html") {
+        "text/html"
+    } else if path.ends_with(".js") {
+        "application/javascript"
+    } else if path.ends_with(".wasm") {
+        "application/wasm"
+    } else if path.ends_with(".css") {
+        "text/css"
+    } else {
+        "application/octet-stream"
+    }
+}
+
+fn start_web_server() {
+    let server = Server::http("0.0.0.0:8080").unwrap();
+    println!("🚀 Serving at http://localhost:8080");
+
+    for request in server.incoming_requests() {
+        let url = request.url();
+        let path = if url == "/" {
+            "static/index.html".to_string()
+        } else {
+            format!("static{}", url)
+        };
+
+        match fs::read(&path) {
+            Ok(data) => {
+                let content_type = mime_type(&path);
+                let response = Response::from_data(data)
+                    .with_header(tiny_http::Header::from_bytes(&b"Content-Type"[..], content_type).unwrap());
+                let _ = request.respond(response);
+            }
+            Err(_) => {
+                let response = Response::from_string("404 Not Found").with_status_code(404);
+                let _ = request.respond(response);
+            }
+        }
+    }
+}
+
+fn launch_expo() {
+    let mut cmd = Command::new("npx");
+    cmd.arg("expo").arg("start").arg("--tunnel");
+    cmd.current_dir("src/native-bridge");
+
+    let status = cmd.status().expect("Failed to launch Expo. Is it installed?");
+    if !status.success() {
+        panic!("Expo failed to launch.");
+    }
+}
+
+
+
+#[pyfunction]
+fn run_game(game: &str, web: bool, react_native: bool) {
+    if web {
+        println!("🌐 Launching '{game}' in web mode...");
+        build_wasm(game);
+        launch_browser();
+        start_web_server();
+    } else if react_native {
+        println!("📱 Launching '{game}' in React Native mode...");
+        build_wasm(game);
+        thread::spawn(start_web_server);
+        launch_expo();
+    } else {
+        println!("🖥️  Launching '{game}' in native mode...");
+        start(game).unwrap();
+    }
+}
+
+#[pymodule]
+fn xos(py: Python, m: &PyModule) -> PyResult<()> {
+    m.add_function(wrap_pyfunction!(run_game, m)?)?;
+    Ok(())
 }
