+++ src/main.rs
@@ -109,99 +109,6 @@ fn main() {
 }
 
 //
-// --- Game Launcher
-//
-
-fn run_game(game: &str, web: bool, react_native: bool) {
-    if web {
-        println!("🌐 Launching '{game}' in web mode...");
-        build_wasm(game);
-        launch_browser();
-        start_web_server();
-    } else if react_native {
-        println!("📱 Launching '{game}' in React Native mode...");
-        build_wasm(game);
-        thread::spawn(start_web_server);
-        launch_expo();
-    } else {
-        println!("🖥️  Launching '{game}' in native mode...");
-        xos::start(game).unwrap();
-    }
-}
-
-//
 // --- Tooling
 //
 
-fn build_wasm(game: &str) {
-    let mut command = Command::new("wasm-pack");
-    command
-        .env("GAME_SELECTION", game)
-        .args(["build", "--target", "web", "--out-dir", "static/pkg"]);
-
-    let status = command.status().expect("Failed to run wasm-pack");
-    if !status.success() {
-        panic!("WASM build failed");
-    }
-
-    println!("✅ WASM built to static/pkg/ with game: {}", game);
-}
-
-fn launch_browser() {
-    thread::spawn(|| {
-        thread::sleep(Duration::from_millis(500));
-        let _ = webbrowser::open("http://localhost:8080");
-    });
-}
-
-fn mime_type(path: &str) -> &'static str {
-    if path.ends_with(".html") {
-        "text/html"
-    } else if path.ends_with(".js") {
-        "application/javascript"
-    } else if path.ends_with(".wasm") {
-        "application/wasm"
-    } else if path.ends_with(".css") {
-        "text/css"
-    } else {
-        "application/octet-stream"
-    }
-}
-
-fn start_web_server() {
-    let server = Server::http("0.0.0.0:8080").unwrap();
-    println!("🚀 Serving at http://localhost:8080");
-
-    for request in server.incoming_requests() {
-        let url = request.url();
-        let path = if url == "/" {
-            "static/index.html".to_string()
-        } else {
-            format!("static{}", url)
-        };
-
-        match fs::read(&path) {
-            Ok(data) => {
-                let content_type = mime_type(&path);
-                let response = Response::from_data(data)
-                    .with_header(tiny_http::Header::from_bytes(&b"Content-Type"[..], content_type).unwrap());
-                let _ = request.respond(response);
-            }
-            Err(_) => {
-                let response = Response::from_string("404 Not Found").with_status_code(404);
-                let _ = request.respond(response);
-            }
-        }
-    }
-}
-
-fn launch_expo() {
-    let mut cmd = Command::new("npx");
-    cmd.arg("expo").arg("start").arg("--tunnel");
-    cmd.current_dir("src/native-bridge");
-
-    let status = cmd.status().expect("Failed to launch Expo. Is it installed?");
-    if !status.success() {
-        panic!("Expo failed to launch.");
-    }
-}
